python/gitpython-test/requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
gitdb==4.0.12
GitPython==3.1.45
idna==3.10
python-dotenv==1.1.1
requests==2.32.5
smmap==5.0.2
urllib3==2.5.0
